+++ internal/handler/api_specs.go
@@ -1,0 +1,339 @@
+package handler
+
+import (
+	"errors"
+	"net/http"
+
+	"github.com/rs/zerolog"
+	"github.com/segmentio/ksuid"
+	"github.com/soltiHQ/control-plane/domain/kind"
+	"github.com/soltiHQ/control-plane/domain/model"
+	"github.com/soltiHQ/control-plane/internal/event"
+	"github.com/soltiHQ/control-plane/internal/service/spec"
+	"github.com/soltiHQ/control-plane/internal/storage"
+	"github.com/soltiHQ/control-plane/internal/storage/inmemory"
+	"github.com/soltiHQ/control-plane/internal/transport/http/responder"
+	"github.com/soltiHQ/control-plane/internal/transport/http/response"
+	"github.com/soltiHQ/control-plane/internal/transport/http/route"
+	"github.com/soltiHQ/control-plane/internal/transport/httpctx"
+	"github.com/soltiHQ/control-plane/internal/uikit/htmx"
+	"github.com/soltiHQ/control-plane/internal/uikit/policy"
+	"github.com/soltiHQ/control-plane/internal/uikit/routepath"
+
+	restv1 "github.com/soltiHQ/control-plane/api/rest/v1"
+	apimapv1 "github.com/soltiHQ/control-plane/internal/transport/http/apimap/v1"
+	contentSpec "github.com/soltiHQ/control-plane/ui/templates/content/spec"
+)
+
+// Specs handles /api/v1/specs.
+//
+// Supported:
+//   - GET  /api/v1/specs
+//   - POST /api/v1/specs
+func (a *API) Specs(w http.ResponseWriter, r *http.Request) {
+	route.Resource(w, r, routepath.ApiSpecs,
+		route.Endpoint{Method: http.MethodGet, Perm: kind.SpecsGet, Fn: a.specList},
+		route.Endpoint{Method: http.MethodPost, Perm: kind.SpecsAdd, Fn: func(w http.ResponseWriter, r *http.Request, m httpctx.RenderMode) {
+			a.specUpsert(w, r, m, "", modeCreate)
+		}},
+	)
+}
+
+// SpecsRouter handles /api/v1/specs/{id} and subroutes.
+//
+// Supported:
+//   - GET    /api/v1/specs/{id}
+//   - PUT    /api/v1/specs/{id}
+//   - DELETE /api/v1/specs/{id}
+//   - POST   /api/v1/specs/{id}/deploy
+//   - GET    /api/v1/specs/{id}/sync
+func (a *API) SpecsRouter(w http.ResponseWriter, r *http.Request) {
+	route.Router(w, r, routepath.ApiSpec,
+		route.Subroute{Action: "", Method: http.MethodGet, Perm: kind.SpecsGet, Fn: a.specDetails},
+		route.Subroute{Action: "", Method: http.MethodPut, Perm: kind.SpecsEdit, Fn: func(w http.ResponseWriter, r *http.Request, m httpctx.RenderMode, id string) {
+			a.specUpsert(w, r, m, id, modeUpdate)
+		}},
+		route.Subroute{Action: "", Method: http.MethodDelete, Perm: kind.SpecsEdit, Fn: a.specDelete},
+		route.Subroute{Action: "deploy", Method: http.MethodPost, Perm: kind.SpecsDeploy, Fn: a.specDeploy},
+		route.Subroute{Action: "sync", Method: http.MethodGet, Perm: kind.SpecsGet, Fn: a.specRollouts},
+	)
+}
+
+func (a *API) specList(w http.ResponseWriter, r *http.Request, mode httpctx.RenderMode) {
+	var (
+		limit  = queryInt(r, "limit", 0)
+		filter storage.SpecFilter
+
+		cursor = r.URL.Query().Get("cursor")
+		q      = r.URL.Query().Get("q")
+	)
+	if q != "" {
+		filter = inmemory.NewSpecFilter().Query(q)
+	}
+
+	res, err := a.specSVC.List(r.Context(), spec.ListQuery{
+		Limit:  limit,
+		Cursor: cursor,
+		Filter: filter,
+	})
+	if err != nil {
+		a.logger.Error().Err(err).Msg("spec list failed")
+		response.Unavailable(w, r, mode)
+		return
+	}
+
+	items := mapSlice(res.Items, apimapv1.Spec)
+	response.OK(w, r, mode, &responder.View{
+		Data: restv1.SpecListResponse{
+			Items:      items,
+			NextCursor: res.NextCursor,
+		},
+		Component: contentSpec.List(res.Items, res.NextCursor, q),
+	})
+}
+
+func (a *API) specDetails(w http.ResponseWriter, r *http.Request, mode httpctx.RenderMode, id string) {
+	ts, err := a.specSVC.Get(r.Context(), id)
+	if err != nil {
+		if errors.Is(err, storage.ErrNotFound) {
+			response.NotFound(w, r, mode)
+			return
+		}
+		a.logger.Error().Err(err).Str("spec", id).Msg("spec get failed")
+		response.Unavailable(w, r, mode)
+		return
+	}
+
+	states, err := a.specSVC.RolloutsBySpec(r.Context(), id, inmemory.NewRolloutFilter().BySpecID(id))
+	if err != nil {
+		a.logger.Error().Err(err).Str("spec", id).Msg("spec rollouts failed")
+		response.Unavailable(w, r, mode)
+		return
+	}
+
+	dto := apimapv1.RolloutSpec(ts, states)
+	response.OK(w, r, mode, &responder.View{
+		Data:      dto,
+		Component: contentSpec.Detail(dto, policy.BuildSpecDetail(a.identity(r))),
+	})
+}
+
+func (a *API) specUpsert(w http.ResponseWriter, r *http.Request, mode httpctx.RenderMode, id string, action upsertMode) {
+	in, err := decodeJSON[restv1.SpecCreateRequest](r)
+	if err != nil {
+		response.BadRequest(w, r, mode)
+		return
+	}
+
+	var ts *model.Spec
+
+	switch action {
+	case modeCreate:
+		if in.Name == "" || in.Slot == "" {
+			response.BadRequest(w, r, mode)
+			return
+		}
+		x, err := model.NewSpec(ksuid.New().String(), in.Name, in.Slot)
+		if err != nil {
+			response.BadRequest(w, r, mode)
+			return
+		}
+		ts = x
+	case modeUpdate:
+		x, err := a.specSVC.Get(r.Context(), id)
+		if err != nil {
+			if errors.Is(err, storage.ErrNotFound) {
+				response.NotFound(w, r, mode)
+				return
+			}
+			a.logger.Error().Err(err).Str("spec", id).Msg("spec get failed")
+			response.Unavailable(w, r, mode)
+			return
+		}
+		if in.Name != "" {
+			x.SetName(in.Name)
+		}
+		if in.Slot != "" {
+			x.SetSlot(in.Slot)
+		}
+		ts = x
+	default:
+		response.BadRequest(w, r, mode)
+		return
+	}
+
+	// Kind
+	if in.KindType != "" {
+		ts.SetKindType(kind.TaskKindType(in.KindType))
+	}
+	if in.KindConfig != nil {
+		ts.SetKindConfig(in.KindConfig)
+	}
+
+	// Lifecycle
+	if in.TimeoutMs > 0 {
+		ts.SetTimeoutMs(in.TimeoutMs)
+	}
+	if in.RestartType != "" {
+		ts.SetRestartType(kind.RestartType(in.RestartType))
+	}
+	if in.IntervalMs > 0 {
+		ts.SetIntervalMs(in.IntervalMs)
+	}
+	if action == modeCreate {
+		ts.SetBackoff(model.BackoffConfig{
+			Jitter:  kind.JitterStrategy(in.Jitter),
+			FirstMs: in.BackoffFirstMs,
+			MaxMs:   in.BackoffMaxMs,
+			Factor:  in.BackoffFactor,
+		})
+	} else if in.Jitter != "" || in.BackoffFirstMs > 0 || in.BackoffMaxMs > 0 || in.BackoffFactor > 0 {
+		b := ts.Backoff()
+		if in.Jitter != "" {
+			b.Jitter = kind.JitterStrategy(in.Jitter)
+		}
+		if in.BackoffFirstMs > 0 {
+			b.FirstMs = in.BackoffFirstMs
+		}
+		if in.BackoffMaxMs > 0 {
+			b.MaxMs = in.BackoffMaxMs
+		}
+		if in.BackoffFactor > 0 {
+			b.Factor = in.BackoffFactor
+		}
+		ts.SetBackoff(b)
+	}
+	if in.Admission != "" {
+		ts.SetAdmission(kind.AdmissionStrategy(in.Admission))
+	}
+
+	// Targets
+	if action == modeCreate {
+		if len(in.Targets) > 0 {
+			ts.SetTargets(in.Targets)
+		}
+		if len(in.TargetLabels) > 0 {
+			ts.SetTargetLabels(in.TargetLabels)
+		}
+		if len(in.RunnerLabels) > 0 {
+			ts.SetRunnerLabels(in.RunnerLabels)
+		}
+	} else {
+		if in.Targets != nil {
+			ts.SetTargets(in.Targets)
+		}
+		if in.TargetLabels != nil {
+			ts.SetTargetLabels(in.TargetLabels)
+		}
+		if in.RunnerLabels != nil {
+			ts.SetRunnerLabels(in.RunnerLabels)
+		}
+	}
+
+	if action == modeCreate {
+		if err := a.specSVC.Create(r.Context(), ts); err != nil {
+			a.logger.Error().Err(err).Msg("spec create failed")
+			response.Unavailable(w, r, mode)
+			return
+		}
+		a.logger.Info().Str("spec", ts.ID()).Str("name", ts.Name()).Msg("spec created")
+		a.hub.Record(event.SpecCreated, event.Payload{ID: ts.ID(), Name: ts.Name()})
+		a.hub.Notify(htmx.SpecUpdate)
+		htmx.Redirect(w, routepath.PageSpecs)
+		response.NoContent(w, r)
+		return
+	}
+
+	if err = a.specSVC.Upsert(r.Context(), ts); err != nil {
+		a.logger.Error().Err(err).Str("spec", id).Msg("spec update failed")
+		response.Unavailable(w, r, mode)
+		return
+	}
+	a.logger.Info().Str("spec", id).Msg("spec updated")
+	a.hub.Record(event.SpecUpdated, event.Payload{ID: id, Name: ts.Name()})
+	htmx.Trigger(w, htmx.SpecUpdate)
+	a.hub.Notify(htmx.SpecUpdate)
+	response.NoContent(w, r)
+}
+
+func (a *API) specDelete(w http.ResponseWriter, r *http.Request, mode httpctx.RenderMode, id string) {
+	err := a.specSVC.Delete(r.Context(), id)
+	if err != nil && !errors.Is(err, storage.ErrNotFound) {
+		a.logger.Error().Err(err).Str("spec", id).Msg("spec delete failed")
+		response.Unavailable(w, r, mode)
+		return
+	}
+	a.logger.Info().Str("spec", id).Msg("spec deleted")
+	a.hub.Notify(htmx.SpecUpdate)
+	htmx.Redirect(w, routepath.PageSpecs)
+	response.NoContent(w, r)
+}
+
+func (a *API) specDeploy(w http.ResponseWriter, r *http.Request, mode httpctx.RenderMode, id string) {
+	if err := a.specSVC.Deploy(r.Context(), id); err != nil {
+		if errors.Is(err, storage.ErrNotFound) {
+			response.NotFound(w, r, mode)
+			return
+		}
+		a.logger.Error().Err(err).Str("spec", id).Msg("spec deploy failed")
+		response.Unavailable(w, r, mode)
+		return
+	}
+
+	rollouts, err := a.specSVC.RolloutsBySpec(r.Context(), id, inmemory.NewRolloutFilter().BySpecID(id))
+	if err != nil {
+		a.logger.Warn().Err(err).Str("spec", id).Msg("spec deployed but rollout query failed")
+	} else {
+		a.logger.Info().Str("spec", id).Int("rollouts", len(rollouts)).Msg("spec deployed")
+		if a.logger.GetLevel() <= zerolog.DebugLevel {
+			for _, ss := range rollouts {
+				a.logger.Debug().
+					Str("rollout_id", ss.ID()).
+					Str("agent_id", ss.AgentID()).
+					Str("status", ss.Status().String()).
+					Int("desired", ss.DesiredVersion()).
+					Int("actual", ss.ActualVersion()).
+					Msg("deploy rollout")
+			}
+		}
+	}
+
+	var specName string
+	if ts, err := a.specSVC.Get(r.Context(), id); err == nil {
+		specName = ts.Name()
+	}
+	a.hub.Record(event.SpecDeployed, event.Payload{ID: id, Name: specName})
+	htmx.Trigger(w, htmx.SpecUpdate)
+	a.hub.Notify(htmx.SpecUpdate)
+	response.NoContent(w, r)
+}
+
+func (a *API) specRollouts(w http.ResponseWriter, r *http.Request, mode httpctx.RenderMode, id string) {
+	states, err := a.specSVC.RolloutsBySpec(r.Context(), id, inmemory.NewRolloutFilter().BySpecID(id))
+	if err != nil {
+		a.logger.Error().Err(err).Str("spec", id).Msg("spec rollouts failed")
+		response.Unavailable(w, r, mode)
+		return
+	}
+
+	a.logger.Info().Str("spec", id).Int("count", len(states)).Msg("spec rollouts loaded")
+	if a.logger.GetLevel() <= zerolog.DebugLevel {
+		for _, ss := range states {
+			a.logger.Debug().
+				Str("rollout_id", ss.ID()).
+				Str("agent_id", ss.AgentID()).
+				Str("status", ss.Status().String()).
+				Int("desired", ss.DesiredVersion()).
+				Int("actual", ss.ActualVersion()).
+				Int("attempts", ss.Attempts()).
+				Str("error", ss.Error()).
+				Msg("rollout entry")
+		}
+	}
+
+	items := mapSlice(states, apimapv1.RolloutEntry)
+	response.OK(w, r, mode, &responder.View{
+		Data:      items,
+		Component: contentSpec.Rollouts(items),
+	})
+}
